+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/web/PurchaseOrderLineController.java
@@ -25,13 +25,10 @@ import com.axelor.exception.AxelorException;
 import com.axelor.inject.Beans;
 import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
-import com.google.inject.Inject;
 import com.google.inject.Singleton;
 
 @Singleton
 public class PurchaseOrderLineController {
-
-  @Inject protected PurchaseOrderLineServiceSupplychainImpl purchaseOrderLineServiceSupplychainImpl;
 
   public void computeAnalyticDistribution(ActionRequest request, ActionResponse response)
       throws AxelorException {
@@ -39,7 +36,8 @@ public class PurchaseOrderLineController {
 
     if (Beans.get(AppAccountService.class).getAppAccount().getManageAnalyticAccounting()) {
       purchaseOrderLine =
-          purchaseOrderLineServiceSupplychainImpl.computeAnalyticDistribution(purchaseOrderLine);
+          Beans.get(PurchaseOrderLineServiceSupplychainImpl.class)
+              .computeAnalyticDistribution(purchaseOrderLine);
       response.setValue("analyticMoveLineList", purchaseOrderLine.getAnalyticMoveLineList());
     }
   }
@@ -49,8 +47,8 @@ public class PurchaseOrderLineController {
     PurchaseOrderLine purchaseOrderLine = request.getContext().asType(PurchaseOrderLine.class);
 
     purchaseOrderLine =
-        purchaseOrderLineServiceSupplychainImpl.createAnalyticDistributionWithTemplate(
-            purchaseOrderLine);
+        Beans.get(PurchaseOrderLineServiceSupplychainImpl.class)
+            .createAnalyticDistributionWithTemplate(purchaseOrderLine);
     response.setValue("analyticMoveLineList", purchaseOrderLine.getAnalyticMoveLineList());
   }
 
@@ -58,8 +56,8 @@ public class PurchaseOrderLineController {
     PurchaseOrderLine purchaseOrderLine = request.getContext().asType(PurchaseOrderLine.class);
     PurchaseOrder purchaseOrder = request.getContext().getParent().asType(PurchaseOrder.class);
 
-    purchaseOrderLineServiceSupplychainImpl.computeBudgetDistributionSumAmount(
-        purchaseOrderLine, purchaseOrder);
+    Beans.get(PurchaseOrderLineServiceSupplychainImpl.class)
+        .computeBudgetDistributionSumAmount(purchaseOrderLine, purchaseOrder);
 
     response.setValue(
         "budgetDistributionSumAmount", purchaseOrderLine.getBudgetDistributionSumAmount());