+++ CHANGELOG.md
@@ -251,5 +251,6 @@
 - TEAMTASK: Add fullname in demo data.
 - Timesheet line : Duration label issue when created from mobile app
 - EXPENSE : fix NPE on validate Expense.
+- Controller: Use Beans.get instead of Inject
 
 [Unreleased 5.2.0]: https://github.com/axelor/axelor-business-suite/compare/dev...wip